[temp dev-pin] protocol → local path for large-file work (restore to exact 1.0.0-rc27 at release)
Local protocol is ahead of rc20: RemoteFile.size is now UInt64 (cast
the FileVault entry size at the fileListReply call site) and the protocol
now ships PrivilegeNames, colliding with HeidrunServerKit's in heidrun-admin
(qualified the refs). Both are mechanical fixes to keep the build green.

+++ Sources/HeidrunServerKit/PacketEncoder.swift
@@ -512,7 +512,7 @@ enum PacketEncoder {
                     name: entry.name,
                     type: entry.type,
                     creator: entry.creator,
-                    size: entry.size,
+                    size: UInt64(entry.size),
                     itemCount: entry.itemCount
                 ),
                 encoding: encoding
